+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [3.5.4] - 2025-09-29
+
+### Fixed
+
+* fixed an issue where `server.json` logs were not picked up from the archive directory
+* updated golang to 1.24.6 for security fixes
+
 ## [3.5.3] - 2025-07-03
 
 ### Added
@@ -938,6 +945,7 @@ someone has added the PAT which is always available
 
 - able to capture logs, configuration and diagnostic data from Dremio clusters deployed on Kubernetes and on-prem
 
+[3.5.4]: https://github.com/dremio/dremio-diagnostic-collector/compare/v3.5.3...v3.5.4
 [3.5.3]: https://github.com/dremio/dremio-diagnostic-collector/compare/v3.5.2...v3.5.3
 [3.5.2]: https://github.com/dremio/dremio-diagnostic-collector/compare/v3.5.1...v3.5.2
 [3.5.1]: https://github.com/dremio/dremio-diagnostic-collector/compare/v3.5.0...v3.5.1